backend/src/requirements.txt
fastapi==0.109.1
uvicorn==0.22.0
PyYAML>=6.0.0,<7.0.0
paho-mqtt>=2.1.0,<3.0.0
python-string-utils==1.0.0
deepmerge==2.0
requests
